requirements-pip.txt
backports.zstd==1.3.0
beautifulsoup4==4.14.3
blinker==1.9.0
Brotli==1.2.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
dotenv==0.9.9
Flask==3.1.2
h2==4.3.0
hpack==4.1.0
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.5.0
pip==25.3
protobuf==3.20.3
PySocks==1.7.1
python-dotenv==1.2.1
requests==2.32.5
soupsieve==2.8.1
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
win_inet_pton==1.1.0
